pyproject.toml
[tool.poetry]
name = "messangers_sender"
version = "0.1.0"
description = ""
authors = ["Suslov Vsevolod <vsevooolod@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
pytz = "^2021.3"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
